+++ src/MonoidK.ts
@@ -3,11 +3,11 @@
  *
  * `MonoidK` instances should satisfy the following laws in addition to the `SemigroupK` laws:
  *
- * 1. Left identity: `zero |> combineK(() => fa) <-> fa`
- * 2. Right identity: `fa |> combineK(() => zero) <-> fa`
- * 3. Annihilation1: `zero |> map(f) <-> zero`
+ * 1. Left identity: `emptyK |> combineK(() => fa) <-> fa`
+ * 2. Right identity: `fa |> combineK(() => emptyK) <-> fa`
+ * 3. Annihilation1: `emptyK |> map(f) <-> emptyK`
  * 4. Distributivity: `fab |> combineK(() => gab) |> ap(fa) <-> fab |> ap(fa) |> combineK(() => gab |> A.ap(fa))`
- * 5. Annihilation2: `zero |> ap(fa) <-> zero`
+ * 5. Annihilation2: `emptyK |> ap(fa) <-> emptyK`
  *
  * @since 3.0.0
  */
@@ -25,7 +25,7 @@ import type { Pointed } from './Pointed'
  * @since 3.0.0
  */
 export interface MonoidK<F extends HKT> extends SemigroupK<F> {
-  readonly zero: <S, R = unknown, W = never, E = never, A = never>() => Kind<F, S, R, W, E, A>
+  readonly emptyK: <S, R = unknown, W = never, E = never, A = never>() => Kind<F, S, R, W, E, A>
 }
 
 // -------------------------------------------------------------------------------------
@@ -39,7 +39,7 @@ export interface MonoidK<F extends HKT> extends SemigroupK<F> {
 export const guard =
   <F extends HKT>(F: MonoidK<F>, P: Pointed<F>) =>
   <S, R = unknown, W = never, E = never>(b: boolean): Kind<F, S, R, W, E, void> =>
-    b ? P.of(undefined) : F.zero()
+    b ? P.of(undefined) : F.emptyK()
 
 // -------------------------------------------------------------------------------------
 // utils
@@ -51,4 +51,4 @@ export const guard =
 export const combineKAll = <F extends HKT>(
   F: MonoidK<F>
 ): (<S, R, W, E, A>(as: ReadonlyArray<Kind<F, S, R, W, E, A>>) => Kind<F, S, R, W, E, A>) =>
-  semigroupK.combineKAll(F)(F.zero())
+  semigroupK.combineKAll(F)(F.emptyK())